refactor(tests): eradicate legacy test suite and scaffold AI-native fuzzing manifolds
Replaces 39 legacy unit tests with 5 high-density property-based testing manifolds using hypothesis. Fixes test typing, unused ignores, and formatting to conform to strict ontology CI/CD rules.
