--- tests/test_cli.py
@@ -1,88 +1,0 @@
-# Copyright (c) 2025 CoReason, Inc.. All Rights Reserved
-#
-# This software is licensed under the Prosperity Public License 3.0.0.
-# The issuer of the Prosperity Public License for this software is CoReason, Inc..
-#
-# For a commercial version of this software, please contact us at gowtham.rao@coreason.ai.
-
-import json
-from pathlib import Path
-from unittest.mock import patch
-
-import pytest
-
-from coreason_manifest.cli.export import main as export_main
-from coreason_manifest.cli.visualize import main as visualize_main
-
-
-def test_export_main(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.chdir(tmp_path)
-    export_main()
-    assert (tmp_path / "coreason_ontology.schema.json").exists()
-
-
-def test_visualize_valid_manifest(tmp_path: Path, capsys: pytest.CaptureFixture[str]) -> None:
-    payload_path = tmp_path / "valid_manifest.json"
-    manifest_data = {
-        "manifest_id": "manifest-test-01",
-        "artifact_profile": {
-            "artifact_event_id": "root-artifact",
-            "detected_modalities": ["text", "tabular_grid"],
-            "token_density": 100,
-        },
-        "active_subgraphs": {"text": ["did:web:agent-1"]},
-        "bypassed_steps": [
-            {
-                "artifact_event_id": "root-artifact",
-                "bypassed_node_id": "did:web:agent-2",
-                "justification": "modality_mismatch",
-                "cryptographic_null_hash": "a" * 64,
-            }
-        ],
-        "branch_budgets_magnitude": {"did:web:agent-1": 1000},
-    }
-    payload_path.write_text(json.dumps(manifest_data))
-
-    with (
-        patch("sys.argv", ["coreason-visualize", str(payload_path)]),
-        patch("sys.exit", side_effect=SystemExit) as mock_exit,
-    ):
-        with pytest.raises(SystemExit):
-            visualize_main()
-        mock_exit.assert_called_once_with(0)
-
-    captured = capsys.readouterr()
-    assert "graph TD" in captured.out
-    assert "did:web:agent-1" in captured.out
-    assert "did:web:agent-2" in captured.out
-    assert ":::active" in captured.out
-    assert ":::bypassed" in captured.out
-
-
-def test_visualize_invalid_manifest(tmp_path: Path, capsys: pytest.CaptureFixture[str]) -> None:
-    payload_path = tmp_path / "invalid_manifest.json"
-    payload_path.write_text('{"invalid": "data"}')
-
-    with (
-        patch("sys.argv", ["coreason-visualize", str(payload_path)]),
-        patch("sys.exit", side_effect=SystemExit) as mock_exit,
-    ):
-        with pytest.raises(SystemExit):
-            visualize_main()
-        mock_exit.assert_called_once_with(1)
-
-    captured = capsys.readouterr()
-    assert "Topological Validation Error" in captured.err
-
-
-def test_visualize_missing_file(capsys: pytest.CaptureFixture[str]) -> None:
-    with (
-        patch("sys.argv", ["coreason-visualize", "ghost_file.json"]),
-        patch("sys.exit", side_effect=SystemExit) as mock_exit,
-    ):
-        with pytest.raises(SystemExit):
-            visualize_main()
-        mock_exit.assert_called_once_with(1)
-
-    captured = capsys.readouterr()
-    assert "not found" in captured.err
